requirements.txt
Django==4.2.19
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
django-simple-history==3.4.0
django-storages==1.14.2



kombu==5.3.5
psycopg2-binary==2.9.9
Pillow==10.1.0
pyjwt==2.8.0
redis==5.0.1
celery==5.3.6
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0

prometheus-client==0.19.0
psutil==5.9.6
djangorestframework-simplejwt==5.2.2